StormsDataset object

Choose the database to use within the package's functions

Details

The fields input must provide at least 6 mandatory fields (and at most 11) in order to benefit from all the functionalities of this package:

A field called "name": which dimension contains the names of storms in the netcdf database

A field called "seasons": which dimension contains the cyclonic seasons of storms in the netcdf database

A field called "isoTime": which dimension contains the ISO times of each (3 or 6 hourly) observations for all storms in the database

A field called "lon": which dimension contains the longitude coordinates (Eastern degree) of each observations for all storms in the netcdf database

A field called "lat": which dimension contains the latitude coordinates (Nothern degree) of each observations for all storms in the netcdf database

A field called "msw": which dimension contains the maximum sustained wind speed (knt) of each observations for all storms in the netcdf database

The following fields are optional but highly recommanded:

A field called "basin": which dimension contains the basin location of storms in the netcdf database. Used to filter the storms in the netcdf database

A field called "rmw": which dimension contains the radius of maximum wind speed (nm) of each observations for all storms in the netcdf database (See stormBehaviour_sp, stormBehaviour_pt)

A field called "sshs": which dimension contains the Saffir Simpson Hurricane Scale index of each observations for all storms in the netcdf database

Finally these following fields are optional but mandatory to perform Holland model (See stormBehaviour_sp, stormBehaviour_pt)

A field called "pressure": which dimension contains the pressure (mb) in the eye for of each observations for all storms in the netcdf database

A field called "poci": which dimension contains the Pressure at the Outermost Closed Isobar (mb) for of each observations for all storms in the nectdf database

Default value is set according to the most relevant dimensions of IBTraCS databases: fields = fields = c("basin" = "basin", "names" = "name", "seasons" = "season", "isoTime" = "iso_time", "lon" = "usa_lon", "lat" = "usa_lat", "msw" = "usa_wind", "rmw" = "usa_rmw", "pressure" = "usa_pres", "poci" = "usa_poci", "sshs" = "usa_sshs")

Slots

filename

character. Name of the database to load. Must be a netcdf file

fields

named character vector. Dictionary that provides all the name of dimensions to extract from the netcdf database (See Details)

basin

character. Basin name to filter the database within its boundaries. Default value is set to NULL. It must be either

"NA": North Atlantic

"SA": South Atlantic

"EP": Eastern North Pacific

"WP": Western North Pacific

"SP": South Pacific

"SI": South India

"NI": North India

database

list of 6 to 10 slots depending on the fields input. Each slot is either a 1D array of dimension (number of storms), for "names" and "seasons" fields, or a 2D array of dimension (Maximum number of observations:number of storms), for the remaining fields which are "isoTime", "lon", "lat", "msw", "rmw", "pressure", "poci", "sshs"