Initialize a StormsDataset object

Initialize a StormsDataset object

Usage

defDatabase(
  filename = system.file("extdata", "IBTrACS.SP.v04r00.nc", package = "StormR"),
  fields = c(basin = "basin", names = "name", seasons = "season", isoTime = "iso_time",
    lon = "usa_lon", lat = "usa_lat", msw = "usa_wind", sshs = "usa_sshs", rmw =
    "usa_rmw", pressure = "usa_pres", poci = "usa_poci"),
  basin = "SP",
  unit_conversion = c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci =
    "mb_to_pa"),
  verbose = TRUE
)

Arguments

filename

character. Name of the database to load. Must be a netcdf file

fields

named character vector. Dictionary that provides all the name of dimension to extract from the netcdf database (See StormsDataSet class)

basin

character. Basin name to filter the database within its boundaries. Default value is set to NULL. It must be either

"NA": North Atlantic

"SA": South Atlantic

"EP": Eastern North Pacific

"WP": Western North Pacific

"SP": South Pacific

"SI": South India

"NI": North India

unit_conversion

...

verbose

logical. Whether or not the function should display informations about the process

Value

An object of class StormsDataset