Getting observations

The getObs() function returns observed track data for a storm in a storm or stormsList object.

Usage

getObs(s, ...)

# S4 method for stormsList
getObs(s, name, season = NULL)

# S4 method for storm
getObs(s)

Arguments

s

storm or stormsList object

...

extra argument for stormsList

name

character. Name of the storm in capital letters.

season

numeric. Cyclonic season of the storm. Required only if several storm in the s object have the same name. Default value is set to NULL.

Value

A data.frame.

Examples

# \donttest{
#Creating a stormsDataset
sds <- defStormsDataset()
#> Warning: No basin argument specified. StormR will work as expected
#>              but cannot use basin filtering for speed-up when collecting data
#> === Loading data  ===
#> Open database... /home/runner/work/_temp/Library/StormR/extdata/test_dataset.nc opened
#> Collecting data ...
#> === DONE ===

#Getting storm track data for all storms near New Caledonia
sts <- defStormsList(sds=sds, loi = "New Caledonia")
#> === Storms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching storms from 2015 to 2021 ...
#>    -> Identifying Storms: 9 potential candidates...
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |=========                                                             |  12%  |                                                                              |==================                                                    |  25%  |                                                                              |==========================                                            |  38%  |                                                                              |===================================                                   |  50%  |                                                                              |============================================                          |  62%  |                                                                              |====================================================                  |  75%  |                                                                              |=============================================================         |  88%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 0.5276096 sec ===
#> 
#> SUMMARY:
#> (*) LOI: New Caledonia 
#> (*) Buffer size: 300 km
#> (*) Number of storms: 7 
#>         Name - Tropical season - Scale - Number of observation within buffer:
#>         PAM - 2015 - 6 - 5 
#>         SOLO - 2015 - 1 - 15 
#>         ULA - 2016 - 5 - 9 
#>         UESI - 2020 - 2 - 16 
#>         GRETEL - 2020 - 2 - 11 
#>         LUCAS - 2021 - 2 - 16 
#>         NIRAN - 2021 - 6 - 10 
#> 

#Getting the observed track data for the tropical
#cyclone Niran in the sts object
getObs(getStorm(sts, name = "NIRAN"))
#>               iso.time      lon       lat msw scale rmw   pres   poci
#> 1  2021-02-27 06:00:00 150.0000 -15.70000  10     0  74 100000 100500
#> 2  2021-02-27 09:00:00 149.6901 -15.84261  10     0  74 100100 100500
#> 3  2021-02-27 12:00:00 149.4000 -16.00000  10     0  74 100200 100500
#> 4  2021-02-27 15:00:00 149.1424 -16.18494  10     0  74 100200 100400
#> 5  2021-02-27 18:00:00 148.9000 -16.40000  10     0  74 100200 100300
#> 6  2021-02-27 21:00:00 148.6201 -16.67255  11     0  74 100100 100300
#> 7  2021-02-28 00:00:00 148.4000 -16.90000  13     0  74 100000 100300
#> 8  2021-02-28 03:00:00 148.3199 -16.97994  14     0  74  99800 100200
#> 9  2021-02-28 06:00:00 148.3000 -17.00000  15     0  74  99600 100100
#> 10 2021-02-28 09:00:00 148.2800 -17.04245  16     0  74  99600 100200
#> 11 2021-02-28 12:00:00 148.2000 -17.10000  18     1  74  99600 100400
#> 12 2021-02-28 15:00:00 147.9726 -17.20757  16     0  74  99600 100400
#> 13 2021-02-28 18:00:00 147.7000 -17.30000  15     0  74  99600 100400
#> 14 2021-02-28 21:00:00 147.4850 -17.35247  16     0  74  99600 100400
#> 15 2021-03-01 00:00:00 147.3000 -17.30000  18     1  74  99600 100400
#> 16 2021-03-01 03:00:00 147.1274 -17.10249  19     1  74  99400 100200
#> 17 2021-03-01 06:00:00 147.0000 -16.80000  21     1  74  99200 100000
#> 18 2021-03-01 09:00:00 146.8974 -16.45746  21     1  69  99300 100100
#> 19 2021-03-01 12:00:00 146.9000 -16.10000  21     1  65  99400 100200
#> 20 2021-03-01 15:00:00 147.0627 -15.77754  22     1  50  98900 100200
#> 21 2021-03-01 18:00:00 147.3000 -15.50000  23     1  37  98400 100200
#> 22 2021-03-01 21:00:00 147.5075 -15.27000  23     1  41  98900 100200
#> 23 2021-03-02 00:00:00 147.7000 -15.10000  23     1  46  99400 100300
#> 24 2021-03-02 03:00:00 147.8725 -14.97750  26     1  46  99100 100300
#> 25 2021-03-02 06:00:00 148.0000 -14.90000  28     1  46  98800 100300
#> 26 2021-03-02 09:00:00 148.1025 -14.84246  28     1  50  98700 100300
#> 27 2021-03-02 12:00:00 148.1000 -14.80000  28     1  56  98600 100400
#> 28 2021-03-02 15:00:00 147.8850 -14.73510  28     1  56  98400 100200
#> 29 2021-03-02 18:00:00 147.7000 -14.70000  28     1  56  98300 100100
#> 30 2021-03-02 21:00:00 147.8275 -14.74254  29     1  56  98300 100200
#> 31 2021-03-03 00:00:00 148.0000 -14.80000  31     1  56  98300 100400
#> 32 2021-03-03 03:00:00 147.9650 -14.79999  33     2  46  97600 100400
#> 33 2021-03-03 06:00:00 147.9000 -14.80000  36     2  37  97000 100400
#> 34 2021-03-03 09:00:00 147.9275 -14.84999  33     2  37  97700 100400
#> 35 2021-03-03 12:00:00 148.0000 -14.90000  31     1  37  98400 100500
#> 36 2021-03-03 15:00:00 148.0776 -14.88498  36     2  30  97500 100300
#> 37 2021-03-03 18:00:00 148.2000 -14.90000  41     2  22  96600 100100
#> 38 2021-03-03 21:00:00 148.3475 -15.02737  42     3  24  96400 100100
#> 39 2021-03-04 00:00:00 148.6000 -15.20000  44     3  28  96200 100100
#> 40 2021-03-04 03:00:00 149.0049 -15.36504  46     3  24  95800 100000
#> 41 2021-03-04 06:00:00 149.5000 -15.50000  49     4  22  95400 100000
#> 42 2021-03-04 09:00:00 149.9179 -15.54925  47     3  22  95700 100100
#> 43 2021-03-04 12:00:00 150.5000 -15.60000  46     3  22  96100 100200
#> 44 2021-03-04 15:00:00 151.4181 -15.69779  47     3  22  95900 100200
#> 45 2021-03-04 18:00:00 152.5000 -15.90000  49     4  22  95800 100200
#> 46 2021-03-04 21:00:00 153.5280 -16.23189  54     4  22  95000 100200
#> 47 2021-03-05 00:00:00 154.6000 -16.70000  59     5  22  94200 100200
#> 48 2021-03-05 03:00:00 155.7347 -17.29210  63     5  15  93400 100000
#> 49 2021-03-05 06:00:00 156.9000 -17.90000  67     5   9  92600  99900
#> 50 2021-03-05 09:00:00 158.0513 -18.35680  69     5   9  92100 100000
#> 51 2021-03-05 12:00:00 159.2000 -18.80000  72     6   9  91700 100100
#> 52 2021-03-05 15:00:00 160.3366 -19.29653  69     5   9  92100 100000
#> 53 2021-03-05 18:00:00 161.5000 -19.90000  67     5   9  92600  99900
#> 54 2021-03-05 21:00:00 162.7202 -20.67662  62     5   9  93500  99900
#> 55 2021-03-06 00:00:00 164.0000 -21.50000  57     4   9  94500  99900
#> 56 2021-03-06 03:00:00 165.3370 -22.19925  55     4   9  94600  99900
#> 57 2021-03-06 06:00:00 166.7000 -22.90000  54     4   9  94800 100000
#> 58 2021-03-06 09:00:00 167.9782 -23.67420  50     4   9  95300 100000
#> 59 2021-03-06 12:00:00 169.4000 -24.50000  46     3   9  95800 100000
#> 60 2021-03-06 15:00:00 171.2547 -25.39050  40     2   9  96700 100100
#> 61 2021-03-06 18:00:00 173.1000 -26.20000  33     2   9  97600 100200
#> 62 2021-03-06 21:00:00 174.4512 -26.75260  31     1  59  98200 100200
#> 63 2021-03-07 00:00:00 175.7000 -27.30000  28     1 111  98800 100300
#> 64 2021-03-07 03:00:00 177.1642 -28.04104  24     1 111  98900 100300
#> 65 2021-03-07 06:00:00 178.8000 -28.90000  21     1 111  99100 100300
getObs(sts, name = "NIRAN")
#>               iso.time      lon       lat msw scale rmw   pres   poci
#> 1  2021-02-27 06:00:00 150.0000 -15.70000  10     0  74 100000 100500
#> 2  2021-02-27 09:00:00 149.6901 -15.84261  10     0  74 100100 100500
#> 3  2021-02-27 12:00:00 149.4000 -16.00000  10     0  74 100200 100500
#> 4  2021-02-27 15:00:00 149.1424 -16.18494  10     0  74 100200 100400
#> 5  2021-02-27 18:00:00 148.9000 -16.40000  10     0  74 100200 100300
#> 6  2021-02-27 21:00:00 148.6201 -16.67255  11     0  74 100100 100300
#> 7  2021-02-28 00:00:00 148.4000 -16.90000  13     0  74 100000 100300
#> 8  2021-02-28 03:00:00 148.3199 -16.97994  14     0  74  99800 100200
#> 9  2021-02-28 06:00:00 148.3000 -17.00000  15     0  74  99600 100100
#> 10 2021-02-28 09:00:00 148.2800 -17.04245  16     0  74  99600 100200
#> 11 2021-02-28 12:00:00 148.2000 -17.10000  18     1  74  99600 100400
#> 12 2021-02-28 15:00:00 147.9726 -17.20757  16     0  74  99600 100400
#> 13 2021-02-28 18:00:00 147.7000 -17.30000  15     0  74  99600 100400
#> 14 2021-02-28 21:00:00 147.4850 -17.35247  16     0  74  99600 100400
#> 15 2021-03-01 00:00:00 147.3000 -17.30000  18     1  74  99600 100400
#> 16 2021-03-01 03:00:00 147.1274 -17.10249  19     1  74  99400 100200
#> 17 2021-03-01 06:00:00 147.0000 -16.80000  21     1  74  99200 100000
#> 18 2021-03-01 09:00:00 146.8974 -16.45746  21     1  69  99300 100100
#> 19 2021-03-01 12:00:00 146.9000 -16.10000  21     1  65  99400 100200
#> 20 2021-03-01 15:00:00 147.0627 -15.77754  22     1  50  98900 100200
#> 21 2021-03-01 18:00:00 147.3000 -15.50000  23     1  37  98400 100200
#> 22 2021-03-01 21:00:00 147.5075 -15.27000  23     1  41  98900 100200
#> 23 2021-03-02 00:00:00 147.7000 -15.10000  23     1  46  99400 100300
#> 24 2021-03-02 03:00:00 147.8725 -14.97750  26     1  46  99100 100300
#> 25 2021-03-02 06:00:00 148.0000 -14.90000  28     1  46  98800 100300
#> 26 2021-03-02 09:00:00 148.1025 -14.84246  28     1  50  98700 100300
#> 27 2021-03-02 12:00:00 148.1000 -14.80000  28     1  56  98600 100400
#> 28 2021-03-02 15:00:00 147.8850 -14.73510  28     1  56  98400 100200
#> 29 2021-03-02 18:00:00 147.7000 -14.70000  28     1  56  98300 100100
#> 30 2021-03-02 21:00:00 147.8275 -14.74254  29     1  56  98300 100200
#> 31 2021-03-03 00:00:00 148.0000 -14.80000  31     1  56  98300 100400
#> 32 2021-03-03 03:00:00 147.9650 -14.79999  33     2  46  97600 100400
#> 33 2021-03-03 06:00:00 147.9000 -14.80000  36     2  37  97000 100400
#> 34 2021-03-03 09:00:00 147.9275 -14.84999  33     2  37  97700 100400
#> 35 2021-03-03 12:00:00 148.0000 -14.90000  31     1  37  98400 100500
#> 36 2021-03-03 15:00:00 148.0776 -14.88498  36     2  30  97500 100300
#> 37 2021-03-03 18:00:00 148.2000 -14.90000  41     2  22  96600 100100
#> 38 2021-03-03 21:00:00 148.3475 -15.02737  42     3  24  96400 100100
#> 39 2021-03-04 00:00:00 148.6000 -15.20000  44     3  28  96200 100100
#> 40 2021-03-04 03:00:00 149.0049 -15.36504  46     3  24  95800 100000
#> 41 2021-03-04 06:00:00 149.5000 -15.50000  49     4  22  95400 100000
#> 42 2021-03-04 09:00:00 149.9179 -15.54925  47     3  22  95700 100100
#> 43 2021-03-04 12:00:00 150.5000 -15.60000  46     3  22  96100 100200
#> 44 2021-03-04 15:00:00 151.4181 -15.69779  47     3  22  95900 100200
#> 45 2021-03-04 18:00:00 152.5000 -15.90000  49     4  22  95800 100200
#> 46 2021-03-04 21:00:00 153.5280 -16.23189  54     4  22  95000 100200
#> 47 2021-03-05 00:00:00 154.6000 -16.70000  59     5  22  94200 100200
#> 48 2021-03-05 03:00:00 155.7347 -17.29210  63     5  15  93400 100000
#> 49 2021-03-05 06:00:00 156.9000 -17.90000  67     5   9  92600  99900
#> 50 2021-03-05 09:00:00 158.0513 -18.35680  69     5   9  92100 100000
#> 51 2021-03-05 12:00:00 159.2000 -18.80000  72     6   9  91700 100100
#> 52 2021-03-05 15:00:00 160.3366 -19.29653  69     5   9  92100 100000
#> 53 2021-03-05 18:00:00 161.5000 -19.90000  67     5   9  92600  99900
#> 54 2021-03-05 21:00:00 162.7202 -20.67662  62     5   9  93500  99900
#> 55 2021-03-06 00:00:00 164.0000 -21.50000  57     4   9  94500  99900
#> 56 2021-03-06 03:00:00 165.3370 -22.19925  55     4   9  94600  99900
#> 57 2021-03-06 06:00:00 166.7000 -22.90000  54     4   9  94800 100000
#> 58 2021-03-06 09:00:00 167.9782 -23.67420  50     4   9  95300 100000
#> 59 2021-03-06 12:00:00 169.4000 -24.50000  46     3   9  95800 100000
#> 60 2021-03-06 15:00:00 171.2547 -25.39050  40     2   9  96700 100100
#> 61 2021-03-06 18:00:00 173.1000 -26.20000  33     2   9  97600 100200
#> 62 2021-03-06 21:00:00 174.4512 -26.75260  31     1  59  98200 100200
#> 63 2021-03-07 00:00:00 175.7000 -27.30000  28     1 111  98800 100300
#> 64 2021-03-07 03:00:00 177.1642 -28.04104  24     1 111  98900 100300
#> 65 2021-03-07 06:00:00 178.8000 -28.90000  21     1 111  99100 100300
# }