Initialize a Storms object

This function returns a Storms object that gathers all the storms specified by the user

getStorms(
  basin = "SP",
  time_period = c(1980, 2022),
  name = NULL,
  loi = NULL,
  max_dist = 300,
  verbose = FALSE,
  remove_TD = TRUE
)

Arguments

basin

character. Name of basin where the storms should be extracted. Default value is set to "SP"

time_period

numeric vector. Should be either one or a range of calendar years. For cyclones that formed in one year and dissipated in the following year, the latter should be used. It could also be a vector of cyclonic season provided that it has the same length as name input and matches the season of each storm listed in name input. Default value is set to c(1980, 2021)

name

character vector. Name(s) of storm(s). Default value is set to NULL, otherwise time_period and name must have the same length, and must match with a storm on the IBTrACS database

loi

Location of Interest. Should be either:

a SpatialPolygon (shapefile)

a sf object

a point of longitude/latitude coordinates provided in a numeric vector

a character representing a country,

a character representing the basin

Default value is set to NULL which will set the spatial.loi.buffer on the whole basin

max_dist

numeric. Indicates the buffer distance (in km) used to generate spatial.loi.buffer. Default value is set to 300km. This value also represents the maximum distance from the track of the storm where computations can be performed

verbose

logical. Whether or not the function must be verbose and display a text progress bar. Default value is set to FALSE

remove_TD

logical. Whether or not to remove tropical depressions (< 18 m/s) and include cyclones only. Default value is set to TRUE.

Value

a Storms object that gathers all storms that match the criteria given in the inputs

Examples

#Focus on a single storm
pam <- getStorms(time_period = 2015, name = "PAM", loi = "Vanuatu")
#> [1] "PAM"

#Focus on several storms over Vanuatu
sts_nc <- getStorms(time_period = c(2003,2021), name = c("ERICA","NIRAN"), loi = "New Caledonia")
#> [1] "ERICA" "NIRAN"

#Focus on every storms that occured in the WP basin between 2010 and 2020
sts_wp <- getStorms(basin = "WP", time_period = c(2010,2020), verbose = TRUE)
#> Making buffer: Done
#> Identifying Storms: Done
#> Loading data: Done
#> Gathering storms 
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
#>   [1] "NIDA"         "OMAIS"        "CONSON"       "CHANTHU"      "DIANMU"      
#>   [6] "MINDULLE"     "LIONROCK"     "KOMPASU"      "NAMTHEUN"     "MALOU"       
#>  [11] "MERANTI"      "FANAPI"       "MALAKAS"      "MEGI"         "CHABA"       
#>  [16] "JAL"          "OMEKA"        "AERE"         "SONGDA"       "HAIMA"       
#>  [21] "MEARI"        "MA-ON"        "NOCK-TEN"     "MUIFA"        "MERBOK"      
#>  [26] "NANMADOL"     "TALAS"        "NORU"         "KULAP"        "ROKE"        
#>  [31] "SONCA"        "NESAT"        "HAITANG"      "NALGAE"       "WASHI"       
#>  [36] "PAKHAR"       "SANVU"        "MAWAR"        "GUCHOL"       "TALIM"       
#>  [41] "DOKSURI"      "KHANUN"       "VICENTE"      "SAOLA"        "DAMREY"      
#>  [46] "HAIKUI"       "KIROGI"       "KAI-TAK"      "TEMBIN"       "BOLAVEN"     
#>  [51] "SANBA"        "JELAWAT"      "EWINIAR"      "GAEMI"        "MALIKSI"     
#>  [56] "PRAPIROON"    "MARIA"        "SON-TINH"     "BOPHA"        "WUKONG"      
#>  [61] "SONAMU"       "YAGI"         "LEEPI"        "BEBINCA"      "RUMBIA"      
#>  [66] "SOULIK"       "CIMARON"      "JEBI"         "MANGKHUT"     "UTOR"        
#>  [71] "TRAMI"        "KONG-REY"     "TORAJI"       "MAN-YI"       "USAGI"       
#>  [76] "PABUK"        "WUTIP"        "FITOW"        "SEPAT"        "DANAS"       
#>  [81] "NARI"         "WIPHA"        "FRANCISCO"    "LEKIMA"       "KROSA"       
#>  [86] "HAIYAN"       "LEHAR"        "KAJIKI"       "FAXAI"        "PEIPAH"      
#>  [91] "TAPAH"        "HAGIBIS"      "NEOGURI"      "RAMMASUN"     "MATMO"       
#>  [96] "HALONG"       "NAKRI"        "FENGSHEN"     "KALMAEGI"     "FUNG-WONG"   
#> [101] "KAMMURI"      "PHANFONE"     "VONGFONG"     "NURI"         "SINLAKU"     
#> [106] "HAGUPIT"      "JANGMI"       "MEKKHALA"     "HIGOS"        "BAVI"        
#> [111] "MAYSAK"       "HAISHEN"      "NOUL"         "DOLPHIN"      "KUJIRA"      
#> [116] "CHAN-HOM"     "LINFA"        "NANGKA"       "SOUDELOR"     "MOLAVE"      
#> [121] "GONI"         "ATSANI"       "ETAU"         "VAMCO"        "KROVANH"     
#> [126] "DUJUAN"       "MUJIGAE"      "CHOI-WAN"     "KOPPU"        "CHAMPI"      
#> [131] "IN-FA"        "MELOR"        "NEPARTAK"     "LUPIT"        "MIRINAE"     
#> [136] "NIDA"         "OMAIS"        "CONSON"       "CHANTHU"      "DIANMU"      
#> [141] "MINDULLE"     "LIONROCK"     "KOMPASU"      "NAMTHEUN"     "MERANTI"     
#> [146] "MALAKAS"      "MEGI"         "CHABA"        "AERE"         "SARIKA"      
#> [151] "HAIMA"        "MEARI"        "MA-ON"        "TOKAGE"       "NOCK-TEN"    
#> [156] "MUIFA"        "MERBOK"       "NANMADOL"     "TALAS"        "NORU"        
#> [161] "KULAP"        "SONCA"        "ROKE"         "NESAT"        "HAITANG"     
#> [166] "NALGAE"       "BANYAN"       "HATO"         "PAKHAR"       "SANVU"       
#> [171] "MAWAR"        "TALIM"        "DOKSURI"      "KHANUN"       "LAN"         
#> [176] "SAOLA"        "DAMREY"       "HAIKUI"       "KIROGI"       "KAI-TAK"     
#> [181] "TEMBIN"       "BOLAVEN"      "SANBA"        "JELAWAT"      "EWINIAR"     
#> [186] "MALIKSI"      "GAEMI"        "PRAPIROON"    "MARIA"        "SON-TINH"    
#> [191] "AMPIL"        "WUKONG"       "JONGDARI"     "SHANSHAN"     "YAGI"        
#> [196] "BEBINCA"      "LEEPI"        "RUMBIA"       "SOULIK"       "CIMARON"     
#> [201] "JEBI"         "MANGKHUT"     "BARIJAT"      "TRAMI"        "KONG-REY"    
#> [206] "YUTU"         "USAGI"        "MAN-YI"       "PABUK"        "WUTIP"       
#> [211] "MUN"          "DANAS"        "NARI"         "WIPHA"        "FRANCISCO"   
#> [216] "LEKIMA"       "KROSA"        "BAILU"        "PODUL"        "KAJIKI"      
#> [221] "LINGLING"     "PEIPAH"       "TAPAH"        "MITAG"        "HAGIBIS"     
#> [226] "NEOGURI"      "BUALOI"       "BULBUL:MATMO" "HALONG"       "NAKRI"       
#> [231] "FENGSHEN"     "KALMAEGI"     "FUNG-WONG"    "KAMMURI"      "PHANFONE"    
#> [236] "VONGFONG"     "NURI"         "SINLAKU"      "HAGUPIT"      "JANGMI"      
#> [241] "MEKKHALA"     "HIGOS"        "BAVI"         "MAYSAK"       "HAISHEN"     
#> [246] "NOUL"         "DOLPHIN"      "KUJIRA"       "CHAN-HOM"     "LINFA"       
#> [251] "NANGKA"       "SAUDEL"       "MOLAVE"       "GONI"         "ATSANI"      
#> [256] "ETAU"         "VAMCO"