Plot rasterized information storm behaviour

This function plots a rasterized product (among maximum sustained wind, power dissipation index, category exposure, or 2D wind speed structure/wind direction at a given observation) associated with a storm provided in a StormsList object alongside with its track

Usage

plotBehaviour(
  sts,
  raster_product,
  color_palette = NULL,
  xlim = NULL,
  ylim = NULL,
  labels = FALSE,
  by = 8,
  pos = 3
)

Arguments

sts

StormsList object

raster_product

SpatRaster object. Name of the layer must be "STORMNAME_product" where product is either:

"MSW"

"PDI"

"Exposure_threshold" where "threshold" represents the wind threshold used to compute Exposure raster

"Speed_index" where index stands for the index of observation

"Direction_index" where index stands for the index of observation

color_palette

character vector. Represents the color palette used for the plot. Default value is set to NULL, which will automatically choose a color palette provided by this package and depending on the product

xlim

numeric vector. A set of longitude coordinates that controls the longitude extent of the plot. Default value is set to NULL which will let the plot extends according to the longitude range of the extended LOI

ylim

numeric vector. A set of latitude coordinates that controls the latitude extent of the plot. Default value is set to NULL which will let the plot extends according to the longitude range of the extended LOI

labels

logical. Whether or not to plot name labels with the corresponding indices of observations and ISO Times along the track

by

numeric. Defines the frequency at which labels are plotted for the 3 (or 6) hourly records. Default value is set to 8 which represents a 24h (or 48h) time interval between each labeled observations. Ignored if labels == FALSE

pos

numeric. Must be between 1 and 4. Correspond to the position of labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right). Default value is set to 3. Ignored if labels == FALSE

Examples

if (FALSE) {
#Plot MSW raster for Pam (2015) in Vanuatu
pam <- Storms(loi = "Vanuatu", names = "PAM")
pam_msw <- spatialBehaviour(pam, verbose = 0)
plotBehaviour(pam, pam_msw)

#Plot a 2D windspeed structure  for Pam (2015) in Vanuatu
pam <- Storms(loi = "Vanuatu", names = "PAM")
pam_prof <- spatialBehaviour(pam, product = "Profiles", verbose = 0)
plotBehaviour(pam, pam_prof[["PAM_Profiles_37"]], labels = TRUE, pos = 2)
}