Compute indicators of storm behaviour

This function computes/rasterizes analytic products for each storm of a Storms object, including Maximum Sustained Wind, Power Dissipation Index, Category exposure and 2D wind speed structures/direction of wind speed for every observations

stormBehaviour_sp(
  sts,
  product = "MSW",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Boose01",
  empirical_rmw = FALSE,
  space_res = "2.5min",
  time_res = 1,
  verbose = 2
)

Arguments

sts

Storms object

product

character. Product to compute among:

"MSW": Maximum Sustained Wind

"PDI": Power Dissipation Index

"Exposure": hour exposition for wind greater than wind_threshold input

"Profiles", 2D wind speed structures of wind speed with wind direction for each observation

Default value is set to "MSW"

wind_threshold

numeric vector. Minimal wind threshold(s) (m/s) to compute Exposure product. Ignored if Exposure is not part of the products to compute. Default value is set to Saffir Simpson Hurricane Scale thresholds

method

character. Cyclonic model used to compute product. Must be either:

"Willoughby": model based on fits performed on cyclonic observations

"Holland": model based on both basic cyclonic Physics and parameters fitting according to cyclonic observations

Default value is set to "Willoughby" (See Details)

asymmetry

character. Indicates which version of asymmetry to use in the computations. Must be either:

"Boose01": formula that substract wind speed to the original wind field depending on the position to the eye of the storm and its velocity

"None": no asymmetry is added

Default value is set to "Boose01" (See Details)

empirical_rmw

logical. Whether to compute the radius of maximum wind empirically or using the radius of maximum wind from the observations. Default value is set to FALSE. If TRUE, a formula extracted from Willoughby et al. 2006 is used to compute rmw

space_res

character. Space resolution for the raster(s) to compute. Either 30sec, 2.5min, 5min or 10min. Default value is set to 2.5min

time_res

numeric. Period of time used to interpolate data. Allowed values are 1 (60min), 0.75 (45min), 0.5 (30min), and 0.25 (15min). Default value is set to 1

verbose

numeric. Whether or not the function should display informations about the process and/or outputs and additional notes. Allowed values are:

0: Nothing is displayed

1: Informations about the process are displayed

2: Outputs are also displayed

3: Additional notes are also displayed

Default value is set to 2

Value

SpatRaster stack which provides the desired product computed, projected in WGS84 and spanning over the extented LOI of the Storms object. Number of layers depends on the number of storm available in sts input and also product and time_res inputs:

"MSW" produces one layer per storm. Name of layer is "STORMNAME_MSW"

"PDI" produces one layer per storm. Name of layer is "STORMNAME_PDI"

"Exposure" produces one layer for each wind values available in wind_threshold input and for each storm. Name of layers are "STORMNAME_Exposure_threshold1", "STORMNAME_Exposure_threshold2"...

"Profiles" produces two layers for each observations (real and interpolated) and each storm. Name of layers are "STORMNAME_Profile_observation", "STORMNAME_WindDirection_observation"

Details

Add details on Willoughy/ Holland method Boose asymmetry and getRMW method ...

Examples

if (FALSE) {
#Compute MSW product for Pam 2015 in Vanuatu using default settings
msw.pam <- stormBehaviour_sp(pam)

#Compute PDI product for Erica and Niran in New Caledonia using Holland model without asymmetry
pdi.nc <- stormBehaviour_sp(sts_nc, method = "Holland", product = "PDI", asymmetry = "None")

#Compute Exposure for Pam 2015 in Vanuatu using default settings
exp.pam <- stormBehaviour_sp(pam, product = "Exposure")

#Compute profiles wind speed for Erica and Niran in New Caledonia using default settings
prof.nc <- stormBehaviour_sp(sts_nc, product = "Profiles")
}