Compute indicators of storm behaviour

This function is the pointwise version of spatialBehaviour. Available products are Time Series of wind speed (TS), Power Dissipation Index (PDI) and Exposure

Usage

temporalBehaviour(
  sts,
  points,
  product = "TS",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Chen",
  empirical_rmw = FALSE,
  temp_res = 1,
  verbose = 1
)

Arguments

sts

StormsList object

points

data.frame. Contains longitude/latitude coordinates within column names "x" and "y", on which to compute the desired product

product

character. Product to compute. Must be either:

"TS": Time Series of wind speed

"PDI": Power Dissipation Index

"Exposure": hour exposition for wind greater than wind_threshold input

Default value is set to "TS"

wind_threshold

numeric vector. Minimal wind threshold(s) (m/s) to compute Exposure product. Ignored if Exposure is not part of the products to compute. Default value is set to Saffir Simpson Hurricane Scale thresholds

method

character. Cyclonic model used to compute product. Must be either:

"Willoughby": model based on fits performed on cyclonic observations

"Holland": model based on both basic cyclonic Physics and parameters fitting according to cyclonic observations

"Boose": asymmetric model based on Holland model

Default value is set to "Willoughby" (See Details)

asymmetry

character. Indicates which version of asymmetry to use in the computations (see Details). Must be either:

"Miyazaki": based on the formula derived in Miyazaki et al. (1962)

"Chen": based on the formula derived in Chen (1994)

"None": no asymmetry is added

Default value is set to "Chen". Ignored if method == Boose

empirical_rmw

logical. Whether to compute the radius of maximum wind empirically or using the radius of maximum wind from the observations. Default value is set to FALSE. If TRUE, a formula extracted from Willoughby et al. 2006 is used to compute rmw

temp_res

numeric. Period of time used to interpolate data. Allowed values are 1 (60min), 0.75 (45min), 0.5 (30min), and 0.25 (15min). Default value is set to 1

verbose

numeric. Whether or not the function should display informations about the process and/or outputs and additional notes. Allowed values are:

0: Nothing is displayed

1: Informations about the process are displayed

Default value is set to 1

Value

Computed product for each points are returned through lists of data.frame (one per points coordinates) contained in a named list. Then, each slot, named after the storm, is made of list(s) of data.frame that has the following dimensions:

If product == "TS": a data frame whose number of rows corresponds to the number of interpolated observations. The columns provides respectively the wind speed values (m/s), the wind directions (in degree), the indices and the ISO time of observations

If product == "PDI": a data.frame with one row and one column that contains PDI value for each point in points.

If product == "Exposure": a data.frame with one row for each wind threshold and one column that contains Exposure value for each point in points.

Examples

if (FALSE) {
pts <- data.frame(lon = c(166.5, 163), lat = c(-22, -19))

#Compute time series of wind speed for ERICA and NIRAN on points
#provided in pts using default settings
ts_nc <- temporalBehaviour(sts_nc, points = pts)

#Compute PDI for ERICA and NIRAN on points provided in pts using default settings
pdiPt_nc <- temporalBehaviour(sts_nc, points = pts, product = "PDI")

#Compute Exposure for ERICA and NIRAN on points provided in df using default settings
expPt_nc <- temporalBehaviour(sts_nc, points = pts, product = "Exposure", wind_threshold = c(20,30))
}