The defDatabase() function creates a StormsDataset object from a NetCDF file

The defDatabase() function creates a StormsDataset object from a NetCDF file

Usage

defDatabase(
  filename,
  fields = c(names = "name", seasons = "season", isoTime = "iso_time", lon = "usa_lon",
    lat = "usa_lat", msw = "usa_wind", sshs = "usa_sshs", rmw = "usa_rmw", pressure =
    "usa_pres", poci = "usa_poci"),
  basin = NULL,
  seasons = c(1980, as.numeric(format(Sys.time(), "%Y"))),
  unit_conversion = c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci =
    "mb_to_pa"),
  verbose = TRUE
)

Arguments

filename

character. Name of the NetCDF (.nc) file.

fields

named character vector. Correspondence between the fields in the output StormsDataset object and the variable names in the input NetCDF file.

basin

character. Name of the basin for which storm track data are extracted. By default basin=NULL, meaning that all stars regardless the basin in which they originated are extracted. Seven basins can be used to filter the data set:

"NA" for North Atlantic basin.

"SA" for South Atlantic basin.

"EP" for Eastern North Pacific basin.

"WP" for Western North Pacific basin.

"SP" for South Pacific basin.

"SI" for South India basin.

"NI" for North India basin.

seasons

numeric vector. Seasons of occurrence of the storms (e.g., c(2020,2022)). By default all storms occurring since 1980 are extracted.

unit_conversion

named character vector. Required unit conversions, msw has to be provided in $m.s^-1$, rmw in $km$, pressure and poci in $Pa$. By default unit_conversion=c(msw = "knt_to_ms", rmw = "nm_to_km", pressure = "mb_to_pa", poci = "mb_to_pa") to meet the conversion requirement when importing a NetCDF file from the IBTrACS database. This argument is mandatory even if no conversion is needed. If no conversion is needed then use "None" in the corresponding fields. The following unit conversions are implemented:

For msw,

"knt_to_ms" to convert knot to meter per second (default setting).

"kmh_to_ms" to convert kilometre per hour to meter per second.

"mph_to_ms" to convert miles per hour to meter per second.

"None"if no conversion is needed.

For rmw,

"nm_to_ms"to convert nautical miles to kilometre (default setting).

"None"if no conversion is needed.

For pressure and poci,

"mb_to_pa" to convert millibar to Pascal (default setting).

"b_to_pa" to convert bar to Pascal.

"atm_to_pa" to convert atmosphere to Pascal.

"psi_to_pa" to convert psi to Pascal.

"None"if no conversion is needed.

verbose

logical. Whether (TRUE) or not (FALSE) the function should display information about the processes

Value

The defDatabase() function returns a StormsDataset object.