Creating a stormsList object

The defStormsList() function extracts storm track data from a stormsDataset and creates a stormsList object based on specified arguments relating to location of interest, seasons, and names of the storms.

Usage

defStormsList(
  sds,
  loi,
  seasons = c(sds@seasons["min"], sds@seasons["max"]),
  names = NULL,
  maxDist = 300,
  removeTD = TRUE,
  verbose = 2
)

Arguments

sds

stormsDataset object.

loi

Location of interest. Can be defined using,

character, a country name (e.g., "Vanuatu")

character, a basin name among "NA", "SA", "EP", "WP", "SP", "SI" and "NI"

numeric vector, a point coordinate (lon, lat in decimal degrees, e.g., c(169.5, -19.2))

sp (SpatialPolygon) or a sf (simple features) object (e.g., created from a shapefile)

seasons

numeric vector. Seasons of occurrence of the storms (e.g., c(2020,2022)). In the Southern Hemisphere, the cyclone season extends across two consecutive years. Therefore, to capture the 2021 to 2022 cyclone season both years should be specified, with cyclones assigned for the year that originated in. By default all storms from sds are extracted.

names

character vector. Names of specific storms (in capital letters).

maxDist

numeric. Maximum distance between the location of interest and the storm for which track data are extracted. Default maxDist is set to 300 km.

removeTD

logical. Whether (TRUE) or not (FALSE) removing tropical depressions (\(msw < 18 m.s^{-1}\), not considered to be stormed in Saffir-Simpson hurricane wind scale) are removed. Default value is set to TRUE.

verbose

numeric. Type of information the function displays. Can be:

2, information about both the processes and the outputs are displayed (default value),

1, only information about the processes are displayed, or

0, nothing is displayed.

Value

The defStormsList() function returns a stormsList object containing track data for all storms meeting the specified criteria (e.g., name, season, location).

Details

The available countries for the loi are those provided in the rwolrdxtra package. This package provide high resolution vector country boundaries derived from Natural Earth data. More informations on the Natural Earth data here: http://www.naturalearthdata.com/downloads/10m-cultural-vectors/.

References

Knapp, K. R., Kruk, M. C., Levinson, D. H., Diamond, H. J., & Neumann, C. J. (2010). The International Best Track Archive for Climate Stewardship (IBTrACS). Bulletin of the American Meteorological Society, 91(3), Article 3. doi:10.1175/2009bams2755.1

Examples

if (FALSE) {
#Creating a stormsDataset
sds <- defStormsDataset()

#Getting data using country names
vanuatu.st <- defStormsList(sds = sds, loi = "Vanuatu")

#Getting data using a specific point location
pt <- c(169, -19)
pam.pt <- defStormsList(sds = sds, loi = pt, names = "PAM")

#Getting data using country and storm names
niran.nc <- defStormsList(sds = sds, loi = "New Caledonia", names = c("NIRAN"))

#Getting data using a user defined spatial polygon
poly <- cbind(c(135, 290, 290, 135, 135),c(-60, -60, 0, 0, -60))
sp <- sf::st_polygon(list(poly))
sp <- sf::st_sfc(sp, crs = 4326)
sp <- sf::st_as_sf(sp)
sts_sp <- defStormsList(sds = sds, loi = sp)
}