Initialize a Storms object

This function returns a Storms object that gathers all the storms specified by the user

getStorms(
  sds = IBTRACS_SP,
  loi,
  seasons = c(1980, max(sds@database$seasons, na.rm = T)),
  names = NULL,
  max_dist = 300,
  remove_TD = TRUE,
  verbose = 2
)

Arguments

sds

StormsDataset object. Default value is set to IBTRACS_SP which is a database provided by this package and based on the IBTrACS.SP.v04r00.nc file. The database must be previously loaded (See Details)

loi

Location of Interest. Must be either:

a SpatialPolygon (shapefile)

a sf object

a point of longitude/latitude coordinates provided in a numeric vector

a character representing a country (See Details section)

seasons

numeric vector. Should be either one or a range of calendar years. For cyclones that formed in one year and dissipated in the following year, the latter should be used

names

character vector. Names of storms. Default value is set to NULL.

max_dist

numeric. Indicates the buffer distance (in km) used to generate spatial.loi.buffer. Default value is set to 300km. This value also represents the maximum distance from the track of the storm where computations can be performed

remove_TD

logical. Whether or not to remove tropical depressions (< 18 m/s) and include cyclones only. Default value is set to TRUE.

verbose

numeric. Either 0, 1, 2 or 3. Whether or not the function must be verbose and display informations about the process, outputs, and additional notes

0 Nothing is displayed

1 Informations about the process are displayed

2 Outputs are also displayed

3 Additional notes to manipulate Storms objects are also displayed

Default value is set to 2

Value

a Storms object that gathers all storms that match the criteria given in the inputs

Examples

#Focus on a single storm
pam <- getStorms(loi = "Vanuatu", names = "PAM")
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching for PAM storm ...
#>    -> Identifying Storms: Done
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 1.617657 sec ===
#> 
#> SUMMARY:
#> (*) LOI: Vanuatu 
#> (*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 1 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         PAM - 2015 - 5 - 20 
#> 

#Focus on several storms over New Caledonia
sts_nc <- getStorms(loi = "New Caledonia", names = c("ERICA","NIRAN"))
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching for ERICA NIRAN storms ...
#>    -> Identifying Storms: Done
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 0.6511176 sec ===
#> 
#> SUMMARY:
#> (*) LOI: New Caledonia 
#> (*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 2 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         ERICA - 2003 - 4 - 19 
#>         NIRAN - 2021 - 5 - 10 
#> 

#Focus on every storms that occured in the SP basin between 2010 and 2020
poly <- cbind(c(135, 290, 290, 135, 135),c(-60, -60, 0, 0, -60))
sp <- sf::st_polygon(list(poly))
sp <- sf::st_sfc(sp, crs = 4326)
sp <- sf::st_as_sf(sp)
sts_sp <- getStorms(loi = sp, seasons = c(2010,2020))
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching storms from 2010 to 2020 ...
#>    -> Identifying Storms: 91 potential candidates...
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 1.182031 sec ===
#> 
#> SUMMARY:
#> (*) LOI: sf object (use getLOI function for further informations(*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 91 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         LAURENCE - 2010 - 4 - 1 
#>         MICK - 2010 - 1 - 35 
#>         OLGA - 2010 - 0 - 69 
#>         NISHA - 2010 - 0 - 33 
#>         OLI - 2010 - 4 - 63 
#>         PAT - 2010 - 2 - 43 
#>         RENE - 2010 - 3 - 77 
#>         SARAH - 2010 - 0 - 93 
#>         TOMAS - 2010 - 4 - 59 
#>         ULUI - 2010 - 5 - 93 
#>         PAUL - 2010 - 1 - 31 
#>         TASHA - 2011 - 0 - 21 
#>         VANIA - 2011 - 0 - 39 
#>         ZELIA - 2011 - 2 - 43 
#>         WILMA - 2011 - 4 - 69 
#>         ANTHONY - 2011 - 0 - 69 
#>         YASI - 2011 - 4 - 66 
#>         ZAKA - 2011 - 0 - 41 
#>         ATU - 2011 - 4 - 67 
#>         BUNE - 2011 - 1 - 53 
#>         JASMINE - 2012 - 4 - 113 
#>         CYRIL - 2012 - 0 - 25 
#>         DAPHNE - 2012 - 0 - 23 
#>         EVAN - 2013 - 4 - 77 
#>         FREDA - 2013 - 3 - 61 
#>         OSWALD - 2013 - 0 - 27 
#>         GARRY - 2013 - 2 - 77 
#>         HALEY - 2013 - 0 - 49 
#>         SANDRA - 2013 - 3 - 95 
#>         TIM - 2013 - 0 - 45 
#>         ZANE - 2013 - 1 - 45 
#>         IAN - 2014 - 4 - 85 
#>         JUNE - 2014 - 0 - 21 
#>         DYLAN - 2014 - 0 - 21 
#>         EDNA - 2014 - 0 - 33 
#>         KOFI - 2014 - 0 - 49 
#>         HADI - 2014 - 0 - 27 
#>         GILLIAN - 2014 - 5 - 87 
#>         LUSI - 2014 - 1 - 47 
#>         MIKE - 2014 - 0 - 19 
#>         ITA - 2014 - 5 - 93 
#>         NIKO - 2015 - 0 - 41 
#>         OLA - 2015 - 2 - 39 
#>         LAM - 2015 - 3 - 35 
#>         MARCIA - 2015 - 4 - 45 
#>         PAM - 2015 - 5 - 57 
#>         NATHAN - 2015 - 3 - 119 
#>         REUBEN - 2015 - 0 - 35 
#>         SOLO - 2015 - 0 - 29 
#>         RAQUEL - 2015 - 0 - 47 
#>         TUNI - 2016 - 0 - 37 
#>         ULA - 2016 - 4 - 119 
#>         VICTOR - 2016 - 2 - 71 
#>         TATIANA - 2016 - 0 - 39 
#>         WINSTON - 2016 - 5 - 151 
#>         YALO - 2016 - 0 - 25 
#>         ZENA - 2016 - 2 - 21 
#>         AMOS - 2016 - 2 - 57 
#>         ALFRED - 2017 - 0 - 43 
#>         BART - 2017 - 0 - 19 
#>         BLANCHE - 2017 - 0 - 8 
#>         DEBBIE - 2017 - 4 - 49 
#>         COOK - 2017 - 2 - 45 
#>         DONNA - 2017 - 4 - 69 
#>         ELLA - 2017 - 1 - 67 
#>         FEHI - 2018 - 0 - 43 
#>         GITA - 2018 - 4 - 103 
#>         HOLA - 2018 - 3 - 55 
#>         LINDA - 2018 - 0 - 37 
#>         NORA - 2018 - 3 - 35 
#>         IRIS - 2018 - 0 - 135 
#>         JOSIE - 2018 - 0 - 43 
#>         KENI - 2018 - 2 - 45 
#>         LIUA - 2019 - 0 - 41 
#>         OWEN - 2019 - 2 - 177 
#>         PENNY - 2019 - 0 - 97 
#>         MONA - 2019 - 0 - 69 
#>         NEIL - 2019 - 0 - 13 
#>         OMA - 2019 - 1 - 141 
#>         POLA - 2019 - 3 - 37 
#>         TREVOR - 2019 - 3 - 89 
#>         ANN - 2019 - 0 - 56 
#>         RITA - 2020 - 1 - 57 
#>         SARAI - 2020 - 1 - 71 
#>         TINO - 2020 - 2 - 39 
#>         UESI - 2020 - 1 - 67 
#>         VICKY - 2020 - 0 - 35 
#>         WASI - 2020 - 0 - 25 
#>         ESTHER - 2020 - 0 - 34 
#>         GRETEL - 2020 - 1 - 27 
#>         HAROLD - 2020 - 5 - 79 
#>