Initialize a Storms object

This function returns a Storms object that gathers all the storms specified by the user

getStorms(
  sds = IBTRACS_SP,
  loi,
  seasons = c(1980, max(sds@database$seasons, na.rm = T)),
  names = NULL,
  max_dist = 300,
  remove_TD = TRUE,
  verbose = 2
)

Arguments

sds

StormsDataset object. Default value is set to IBTRACS_SP which is a database provided by this package and based on the IBTrACS.SP.v04r00.nc file. (See StormsDataset class)

loi

Location of Interest. Must be either:

a SpatialPolygon (shapefile)

a sf object

a point of longitude/latitude coordinates provided in a numeric vector

a character representing a country (See Details section)

seasons

numeric vector. Should be either one or a range of calendar years. For cyclones that formed in one year and dissipated in the following year, the latter should be used. Default values will allow searching for storms between 1980 and the maximum cyclonic season available in the StormsDataSet object

names

character vector. Names of storms in capital letters. Default value is set to NULL.

max_dist

numeric. Indicates the buffer distance (in km) used to extend the location of interest. Default value is set to 300km

remove_TD

logical. Whether or not to remove tropical depressions (< 18 m/s) and include cyclones only. Default value is set to TRUE

verbose

numeric. Whether or not the function should display informations about the process, outputs, and additional notes. Allowed values are:

0: Nothing is displayed

1: Informations about the process are displayed

2: Outputs are also displayed

3: Additional notes to manipulate Storms objects are also displayed

Default value is set to 2

Value

a Storms object that gathers all storms that match the criteria given in the inputs

Details

Available countries for the loi input are those provided in the rwolrdxtra package, derived from the newest version of Natural Earth Data. Check the following link for more informations: http://www.naturalearthdata.com/downloads/10m-cultural-vectors/

Examples

if (FALSE) {
#Get data for a single storm on a given country
pam <- getStorms(loi = "Vanuatu", names = "PAM")

#Get data for the above storm on a point coordinates (Eastern/Northern degree)
pt <- c(169, -19)
pam.pt <- getStorms(loi = pt, names = "PAM")

#Get data for two storms over New Caledonia
sts_nc <- getStorms(loi = "New Caledonia", names = c("ERICA","NIRAN"))

#Get data for every storms that occured in the SP basin between 2010 and 2020
poly <- cbind(c(135, 290, 290, 135, 135),c(-60, -60, 0, 0, -60))
sp <- sf::st_polygon(list(poly))
sp <- sf::st_sfc(sp, crs = 4326)
sp <- sf::st_as_sf(sp)
sts_sp <- getStorms(loi = sp, seasons = c(2010,2020))
}