Initialize a Storms object

This function returns a Storms object that gathers all the storms specified by the user

getStorms(
  sdb = IBTRACS,
  loi,
  seasons = c(1980, as.numeric(format(Sys.time(), "%Y"))),
  names = NULL,
  max_dist = 300,
  remove_TD = TRUE,
  verbose = 2
)

Arguments

sdb

Storms data base returned by loadData function. Default value is set to IBTRACS which is a data base provided by this package and based on the IBTrACS.All.v04r00.nc file

loi

Location of Interest. Must be either:

a SpatialPolygon (shapefile)

a sf object

a point of longitude/latitude coordinates provided in a numeric vector

a character representing a country (See Details section)

seasons

numeric vector. Should be either one or a range of calendar years. For cyclones that formed in one year and dissipated in the following year, the latter should be used

names

character vector. Names of storms. Default value is set to NULL.

max_dist

numeric. Indicates the buffer distance (in km) used to generate spatial.loi.buffer. Default value is set to 300km. This value also represents the maximum distance from the track of the storm where computations can be performed

remove_TD

logical. Whether or not to remove tropical depressions (< 18 m/s) and include cyclones only. Default value is set to TRUE.

verbose

numeric. Either 0, 1, 2 or 3. Whether or not the function must be verbose and display informations about the process, outputs, and additional notes

0 Nothing is displayed

1 Informations about the process are displayed

2 Outputs are also displayed

3 Additional notes to manipulate Storms objects are also displayed

Default value is set to 2

Value

a Storms object that gathers all storms that match the criteria given in the inputs

Examples

#Focus on a single storm
pam <- getStorms(loi = "Vanuatu", names = "PAM")
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching for PAM storm ...
#>    -> Identifying Storms: Done
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 4.242261 sec ===
#> 
#> SUMMARY:
#> (*) LOI: Vanuatu 
#> (*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 1 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         PAM - 2015 - 5 - 20 
#> 

#Focus on several storms over New Caledonia
sts_nc <- getStorms(loi = "New Caledonia", names = c("ERICA","NIRAN"))
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching for ERICA NIRAN storms ...
#>    -> Identifying Storms: Done
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 0.783617 sec ===
#> 
#> SUMMARY:
#> (*) LOI: New Caledonia 
#> (*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 2 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         ERICA - 2003 - 4 - 19 
#>         NIRAN - 2021 - 5 - 10 
#> 

#Focus on every storms that occured in the WP basin between 2010 and 2020
poly <- cbind(c(100, 180, 180, 100, 100),c(0, 0, 60, 60, 0))
wp <- sf::st_polygon(list(poly))
wp <- sf::st_sfc(wp, crs = 4326)
wp <- sf::st_as_sf(wp)
sts_wp <- getStorms(loi = wp, seasons = c(2010,2020))
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching storms from 2010 to 2020 ...
#>    -> Identifying Storms: 916 potential candidates...
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |=========                                                             |  14%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |=======================                                               |  34%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |===================================================                   |  74%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |=================================================================     |  94%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 13.54063 sec ===
#> 
#> SUMMARY:
#> (*) LOI: sf object (use getLOI function for further informations(*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 271 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         NIDA - 2009 - 5 - 99 
#>         OMAIS - 2010 - 0 - 57 
#>         CONSON - 2010 - 1 - 63 
#>         CHANTHU - 2010 - 1 - 53 
#>         DIANMU - 2010 - 0 - 53 
#>         MINDULLE - 2010 - 1 - 35 
#>         LIONROCK - 2010 - 0 - 61 
#>         KOMPASU - 2010 - 3 - 53 
#>         NAMTHEUN - 2010 - 0 - 23 
#>         MALOU - 2010 - 0 - 67 
#>         MERANTI - 2010 - 1 - 45 
#>         FANAPI - 2010 - 3 - 61 
#>         MALAKAS - 2010 - 2 - 47 
#>         MEGI - 2010 - 5 - 103 
#>         CHABA - 2010 - 4 - 83 
#>         GIRI - 2010 - 4 - 3 
#>         JAL - 2010 - 0 - 14 
#>         OMEKA - 2010 - 0 - 37 
#>         AERE - 2011 - 0 - 53 
#>         SONGDA - 2011 - 5 - 77 
#>         HAIMA - 2011 - 0 - 77 
#>         MEARI - 2011 - 0 - 53 
#>         MA-ON - 2011 - 4 - 93 
#>         NOCK-TEN - 2011 - 1 - 53 
#>         MUIFA - 2011 - 5 - 107 
#>         MERBOK - 2011 - 1 - 55 
#>         NANMADOL - 2011 - 5 - 67 
#>         TALAS - 2011 - 0 - 91 
#>         NORU - 2011 - 0 - 43 
#>         KULAP - 2011 - 0 - 37 
#>         ROKE - 2011 - 4 - 95 
#>         SONCA - 2011 - 2 - 59 
#>         NESAT - 2011 - 4 - 63 
#>         HAITANG - 2011 - 0 - 27 
#>         NALGAE - 2011 - 4 - 71 
#>         WASHI - 2011 - 0 - 61 
#>         PAKHAR - 2012 - 0 - 41 
#>         SANVU - 2012 - 1 - 61 
#>         MAWAR - 2012 - 3 - 51 
#>         GUCHOL - 2012 - 4 - 81 
#>         TALIM - 2012 - 0 - 31 
#>         DOKSURI - 2012 - 0 - 37 
#>         KHANUN - 2012 - 0 - 45 
#>         VICENTE - 2012 - 4 - 39 
#>         SAOLA - 2012 - 2 - 61 
#>         DAMREY - 2012 - 1 - 47 
#>         HAIKUI - 2012 - 1 - 67 
#>         KIROGI - 2012 - 0 - 53 
#>         KAI-TAK - 2012 - 1 - 53 
#>         TEMBIN - 2012 - 4 - 97 
#>         BOLAVEN - 2012 - 4 - 85 
#>         SANBA - 2012 - 5 - 63 
#>         JELAWAT - 2012 - 5 - 99 
#>         EWINIAR - 2012 - 0 - 57 
#>         GAEMI - 2012 - 0 - 55 
#>         MALIKSI - 2012 - 0 - 43 
#>         PRAPIROON - 2012 - 3 - 103 
#>         MARIA - 2012 - 0 - 57 
#>         SON-TINH - 2012 - 3 - 51 
#>         BOPHA - 2012 - 5 - 103 
#>         WUKONG - 2012 - 0 - 41 
#>         SONAMU - 2013 - 0 - 57 
#>         YAGI - 2013 - 0 - 63 
#>         LEEPI - 2013 - 0 - 39 
#>         BEBINCA - 2013 - 0 - 35 
#>         RUMBIA - 2013 - 1 - 39 
#>         SOULIK - 2013 - 4 - 61 
#>         CIMARON - 2013 - 0 - 35 
#>         JEBI - 2013 - 0 - 33 
#>         MANGKHUT - 2013 - 0 - 27 
#>         UTOR - 2013 - 4 - 59 
#>         UNALA - 2013 - 0 - 4 
#>         PEWA - 2013 - 1 - 62 
#>         TRAMI - 2013 - 1 - 57 
#>         KONG-REY - 2013 - 0 - 49 
#>         TORAJI - 2013 - 0 - 29 
#>         MAN-YI - 2013 - 0 - 37 
#>         USAGI - 2013 - 4 - 58 
#>         PABUK - 2013 - 2 - 51 
#>         WUTIP - 2013 - 3 - 41 
#>         FITOW - 2013 - 2 - 59 
#>         SEPAT - 2013 - 0 - 29 
#>         DANAS - 2013 - 4 - 49 
#>         PHAILIN - 2013 - 5 - 1 
#>         NARI - 2013 - 3 - 57 
#>         WIPHA - 2013 - 4 - 55 
#>         FRANCISCO - 2013 - 5 - 85 
#>         LEKIMA - 2013 - 5 - 59 
#>         KROSA - 2013 - 3 - 63 
#>         HAIYAN - 2013 - 5 - 73 
#>         LEHAR - 2013 - 1 - 29 
#>         KAJIKI - 2014 - 0 - 23 
#>         FAXAI - 2014 - 1 - 63 
#>         PEIPAH - 2014 - 0 - 21 
#>         TAPAH - 2014 - 1 - 47 
#>         HAGIBIS - 2014 - 0 - 35 
#>         NEOGURI - 2014 - 5 - 69 
#>         RAMMASUN - 2014 - 5 - 85 
#>         MATMO - 2014 - 2 - 59 
#>         GENEVIEVE - 2014 - 5 - 53 
#>         HALONG - 2014 - 5 - 111 
#>         NAKRI - 2014 - 0 - 59 
#>         FENGSHEN - 2014 - 1 - 35 
#>         KALMAEGI - 2014 - 1 - 57 
#>         FUNG-WONG - 2014 - 0 - 51 
#>         KAMMURI - 2014 - 0 - 45 
#>         PHANFONE - 2014 - 4 - 71 
#>         VONGFONG - 2014 - 5 - 97 
#>         NURI - 2014 - 5 - 61 
#>         SINLAKU - 2014 - 0 - 39 
#>         HAGUPIT - 2014 - 5 - 99 
#>         JANGMI - 2014 - 0 - 37 
#>         MEKKHALA - 2015 - 1 - 63 
#>         HIGOS - 2015 - 4 - 43 
#>         BAVI - 2015 - 0 - 71 
#>         MAYSAK - 2015 - 5 - 85 
#>         HAISHEN - 2015 - 0 - 39 
#>         NOUL - 2015 - 5 - 77 
#>         DOLPHIN - 2015 - 5 - 111 
#>         KUJIRA - 2015 - 0 - 37 
#>         CHAN-HOM - 2015 - 4 - 103 
#>         LINFA - 2015 - 1 - 63 
#>         NANGKA - 2015 - 4 - 121 
#>         HALOLA - 2015 - 2 - 116 
#>         IUNE - 2015 - 0 - 2 
#>         SOUDELOR - 2015 - 5 - 83 
#>         MOLAVE - 2015 - 0 - 59 
#>         GONI - 2015 - 4 - 99 
#>         ATSANI - 2015 - 5 - 89 
#>         KILO - 2015 - 4 - 97 
#>         LOKE - 2015 - 1 - 47 
#>         ETAU - 2015 - 0 - 33 
#>         VAMCO - 2015 - 0 - 17 
#>         KROVANH - 2015 - 3 - 57 
#>         DUJUAN - 2015 - 4 - 73 
#>         MUJIGAE - 2015 - 4 - 43 
#>         CHOI-WAN - 2015 - 1 - 39 
#>         KOPPU - 2015 - 4 - 65 
#>         CHAMPI - 2015 - 4 - 97 
#>         IN-FA - 2015 - 4 - 79 
#>         MELOR - 2015 - 4 - 47 
#>         NEPARTAK - 2016 - 5 - 55 
#>         LUPIT - 2016 - 0 - 13 
#>         MIRINAE - 2016 - 1 - 29 
#>         NIDA - 2016 - 1 - 33 
#>         OMAIS - 2016 - 1 - 41 
#>         CONSON - 2016 - 0 - 55 
#>         CHANTHU - 2016 - 0 - 41 
#>         DIANMU - 2016 - 0 - 21 
#>         MINDULLE - 2016 - 1 - 53 
#>         LIONROCK - 2016 - 4 - 111 
#>         KOMPASU - 2016 - 0 - 21 
#>         NAMTHEUN - 2016 - 3 - 45 
#>         MERANTI - 2016 - 5 - 55 
#>         MALAKAS - 2016 - 4 - 73 
#>         MEGI - 2016 - 4 - 51 
#>         CHABA - 2016 - 5 - 78 
#>         SONGDA - 2016 - 4 - 75 
#>         AERE - 2016 - 0 - 77 
#>         SARIKA - 2016 - 4 - 49 
#>         HAIMA - 2016 - 5 - 59 
#>         MEARI - 2016 - 2 - 45 
#>         MA-ON - 2016 - 0 - 25 
#>         TOKAGE - 2016 - 1 - 39 
#>         NOCK-TEN - 2016 - 5 - 65 
#>         MUIFA - 2017 - 0 - 29 
#>         MERBOK - 2017 - 0 - 25 
#>         NANMADOL - 2017 - 1 - 31 
#>         TALAS - 2017 - 0 - 21 
#>         NORU - 2017 - 4 - 149 
#>         KULAP - 2017 - 0 - 59 
#>         SONCA - 2017 - 0 - 45 
#>         ROKE - 2017 - 0 - 17 
#>         NESAT - 2017 - 2 - 41 
#>         HAITANG - 2017 - 0 - 27 
#>         NALGAE - 2017 - 0 - 47 
#>         BANYAN - 2017 - 3 - 55 
#>         HATO - 2017 - 3 - 35 
#>         PAKHAR - 2017 - 0 - 29 
#>         SANVU - 2017 - 2 - 47 
#>         MAWAR - 2017 - 0 - 23 
#>         TALIM - 2017 - 4 - 71 
#>         DOKSURI - 2017 - 2 - 43 
#>         KHANUN - 2017 - 2 - 33 
#>         LAN - 2017 - 4 - 65 
#>         SAOLA - 2017 - 1 - 83 
#>         DAMREY - 2017 - 2 - 31 
#>         HAIKUI - 2017 - 0 - 53 
#>         KIROGI - 2017 - 0 - 23 
#>         KAI-TAK - 2017 - 0 - 85 
#>         TEMBIN - 2017 - 2 - 49 
#>         BOLAVEN - 2018 - 0 - 29 
#>         SANBA - 2018 - 0 - 61 
#>         JELAWAT - 2018 - 4 - 83 
#>         EWINIAR - 2018 - 0 - 51 
#>         MALIKSI - 2018 - 0 - 39 
#>         GAEMI - 2018 - 0 - 27 
#>         PRAPIROON - 2018 - 1 - 61 
#>         MARIA - 2018 - 5 - 71 
#>         SON-TINH - 2018 - 0 - 87 
#>         AMPIL - 2018 - 0 - 49 
#>         WUKONG - 2018 - 1 - 49 
#>         JONGDARI - 2018 - 2 - 99 
#>         HECTOR - 2018 - 4 - 30 
#>         SHANSHAN - 2018 - 2 - 67 
#>         YAGI - 2018 - 0 - 67 
#>         BEBINCA - 2018 - 0 - 63 
#>         LEEPI - 2018 - 1 - 47 
#>         RUMBIA - 2018 - 0 - 35 
#>         SOULIK - 2018 - 3 - 85 
#>         CIMARON - 2018 - 4 - 63 
#>         JEBI - 2018 - 5 - 71 
#>         MANGKHUT - 2018 - 5 - 87 
#>         BARIJAT - 2018 - 0 - 43 
#>         TRAMI - 2018 - 5 - 87 
#>         KONG-REY - 2018 - 5 - 77 
#>         YUTU - 2018 - 5 - 109 
#>         USAGI - 2018 - 2 - 157 
#>         MAN-YI - 2018 - 2 - 63 
#>         PABUK - 2018 - 0 - 49 
#>         WUTIP - 2019 - 5 - 85 
#>         MUN - 2019 - 0 - 27 
#>         DANAS - 2019 - 0 - 57 
#>         NARI - 2019 - 0 - 29 
#>         ERICK - 2019 - 4 - 6 
#>         WIPHA - 2019 - 0 - 47 
#>         FRANCISCO - 2019 - 1 - 57 
#>         LEKIMA - 2019 - 4 - 67 
#>         KROSA - 2019 - 3 - 99 
#>         BAILU - 2019 - 0 - 57 
#>         PODUL - 2019 - 0 - 53 
#>         FAXAI - 2019 - 4 - 93 
#>         KAJIKI - 2019 - 0 - 27 
#>         LINGLING - 2019 - 4 - 67 
#>         PEIPAH - 2019 - 0 - 41 
#>         TAPAH - 2019 - 1 - 51 
#>         MITAG - 2019 - 2 - 71 
#>         HAGIBIS - 2019 - 5 - 73 
#>         NEOGURI - 2019 - 2 - 57 
#>         BUALOI - 2019 - 5 - 67 
#>         BULBUL:MATMO - 2019 - 3 - 44 
#>         HALONG - 2019 - 5 - 71 
#>         NAKRI - 2019 - 1 - 51 
#>         FENGSHEN - 2019 - 4 - 71 
#>         KALMAEGI - 2019 - 2 - 89 
#>         FUNG-WONG - 2019 - 1 - 39 
#>         KAMMURI - 2019 - 4 - 85 
#>         PHANFONE - 2019 - 3 - 79 
#>         VONGFONG - 2020 - 3 - 73 
#>         NURI - 2020 - 0 - 53 
#>         DOUGLAS - 2020 - 4 - 2 
#>         SINLAKU - 2020 - 0 - 39 
#>         HAGUPIT - 2020 - 1 - 47 
#>         JANGMI - 2020 - 0 - 25 
#>         MEKKHALA - 2020 - 1 - 29 
#>         HIGOS - 2020 - 0 - 27 
#>         BAVI - 2020 - 3 - 53 
#>         MAYSAK - 2020 - 4 - 69 
#>         HAISHEN - 2020 - 4 - 61 
#>         NOUL - 2020 - 0 - 39 
#>         DOLPHIN - 2020 - 0 - 41 
#>         KUJIRA - 2020 - 1 - 35 
#>         CHAN-HOM - 2020 - 1 - 81 
#>         LINFA - 2020 - 0 - 23 
#>         NANGKA - 2020 - 0 - 29 
#>         SAUDEL - 2020 - 1 - 79 
#>         MOLAVE - 2020 - 3 - 59 
#>         GONI - 2020 - 5 - 97 
#>         ATSANI - 2020 - 0 - 55 
#>         ETAU - 2020 - 0 - 35 
#>         VAMCO - 2020 - 4 - 63 
#>