Plot rasterized information storm behaviour

This function plots a rasterize product (maximum sustained wind, power dissipation index, category exposure, or 2D wind speed structure and Wind direction at a given observation) associated with a storm contained in a Storms object alongside with its track

plotBehaviour(
  sts,
  raster_product,
  color_palette = NULL,
  xlim = NULL,
  ylim = NULL,
  labels = FALSE,
  by = 8,
  pos = 3
)

Arguments

sts

Storms object

raster_product

Spatraster object. Name of the layer must be "stormName_product" where product is either "MSW", "PDI", "Exposure_min-max", "Profiles_index", "WindDirection_index" where "min" and "max" represent the threshold used to compute Exposure raster and index stands for the index of observation.

color_palette

character vector. Represents the color palette used for the plot. Default value is set to NULL, which will automatically choose a color palette provided by this package and depending on the product

xlim

numeric vector. A set of longitude coordinates that controls the longitude extent of the plot. Default value is set to NULL which will let the plot extends according to the x bounding box of the spatial LOI buffer provided in sts

ylim

numeric vector. A set of latitude coordinates that controls the latitude extent of the plot. Default value is set to NULL which will let the plot extends according to the y bounding box of the spatial LOI buffer provided in sts

labels

logical. Whether or not to plot ISO Times and name labels

by

numeric. Defines the frequency at which labels are plotted for the 3-hourly records. Default value is set to 8 which represents a 24h time interval between each labeled observations. Ignored if labels == FALSE

pos

numeric. Must be between 1 and 4. Correspond to the position of labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right). Default value is set to 3. Ignored if labels == FALSE

Examples

#Plot MSW analytic raster for Pam (2015) in Vanuatu
pam <- getStorms(loi = "Vanuatu", names = "PAM")
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching for PAM storm ...
#>    -> Identifying Storms: Done
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 0.6978354 sec ===
#> 
#> SUMMARY:
#> (*) LOI: Vanuatu 
#> (*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 1 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         PAM - 2015 - 5 - 20 
#> 
pam_msw <- stormBehaviour_sp(pam)
#> === stormBehaviour_sp processing ... ===
#> 
#> Computation settings:
#>   (*) Output format: analytical rasters (Integration in space and/or time)
#>   (*) Time interpolation: Every 60 min
#>   (*) Space resolution: 2.5min 
#>   (*) Method used: Willoughby 
#>   (*) Product(s) to compute: MSW 
#>   (*) Asymmetry used: Boose01 
#> 
#> Storm(s):
#>   ( 1 )  PAM 
#> 
#> PAM  ( 1 / 1 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |===================                                                   |  26%  |                                                                              |====================                                                  |  28%  |                                                                              |=====================                                                 |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  32%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  35%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================                |  76%  |                                                                              |=======================================================               |  78%  |                                                                              |========================================================              |  79%  |                                                                              |=========================================================             |  81%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 13.701 sec ===
#> 
#> Output:
#> SpatRaster stack with 1 layers:
#> index - names of layer
#>   1     PAM_MSW 
#> 
plotBehaviour(pam, pam_msw)

#Plot a 2D windspeed structure  for Pam (2015) in Vanuatu
pam <- getStorms(loi = "Vanuatu", names = "PAM")
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching for PAM storm ...
#>    -> Identifying Storms: Done
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 0.6832881 sec ===
#> 
#> SUMMARY:
#> (*) LOI: Vanuatu 
#> (*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 1 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         PAM - 2015 - 5 - 20 
#> 
pam_prof <- stormBehaviour_sp(pam, format = "profiles", verbose = 3)
#> === stormBehaviour_sp processing ... ===
#> 
#> Computation settings:
#>   (*) Output format: 2D WindSpeed structures at each observation (real and interpolated)
#>   (*) Space resolution: 2.5min 
#>   (*) Method used: Willoughby 
#>   (*) Asymmetry used: Boose01 
#> 
#> Storm(s):
#>   ( 1 )  PAM 
#> 
#> PAM  ( 1 / 1 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  12%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |================                                                      |  23%  |                                                                              |==================                                                    |  25%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  32%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  39%  |                                                                              |=============================                                         |  41%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  46%  |                                                                              |==================================                                    |  48%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  52%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  55%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  61%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  66%  |                                                                              |================================================                      |  68%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  95%  |                                                                              |====================================================================  |  96%  |                                                                              |===================================================================== |  98%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 9.732 sec ===
#> 
#> Output:
#> SpatRaster stack with 114 layers:
#> index - names of layer
#>   1     PAM_Profiles_28 
#>   2     PAM_Profiles_28.1 
#>   3     PAM_Profiles_28.2 
#>   4     PAM_Profiles_29 
#>   5     PAM_Profiles_29.1 
#>   6     PAM_Profiles_29.2 
#>   7     PAM_Profiles_30 
#>   8     PAM_Profiles_30.1 
#>   9     PAM_Profiles_30.2 
#>   10     PAM_Profiles_31 
#>   11     PAM_Profiles_31.1 
#>   12     PAM_Profiles_31.2 
#>   13     PAM_Profiles_32 
#>   14     PAM_Profiles_32.1 
#>   15     PAM_Profiles_32.2 
#>   16     PAM_Profiles_33 
#>   17     PAM_Profiles_33.1 
#>   18     PAM_Profiles_33.2 
#>   19     PAM_Profiles_34 
#>   20     PAM_Profiles_34.1 
#>   21     PAM_Profiles_34.2 
#>   22     PAM_Profiles_35 
#>   23     PAM_Profiles_35.1 
#>   24     PAM_Profiles_35.2 
#>   25     PAM_Profiles_36 
#>   26     PAM_Profiles_36.1 
#>   27     PAM_Profiles_36.2 
#>   28     PAM_Profiles_37 
#>   29     PAM_Profiles_37.1 
#>   30     PAM_Profiles_37.2 
#>   31     PAM_Profiles_38 
#>   32     PAM_Profiles_38.1 
#>   33     PAM_Profiles_38.2 
#>   34     PAM_Profiles_39 
#>   35     PAM_Profiles_39.1 
#>   36     PAM_Profiles_39.2 
#>   37     PAM_Profiles_40 
#>   38     PAM_Profiles_40.1 
#>   39     PAM_Profiles_40.2 
#>   40     PAM_Profiles_41 
#>   41     PAM_Profiles_41.1 
#>   42     PAM_Profiles_41.2 
#>   43     PAM_Profiles_42 
#>   44     PAM_Profiles_42.1 
#>   45     PAM_Profiles_42.2 
#>   46     PAM_Profiles_43 
#>   47     PAM_Profiles_43.1 
#>   48     PAM_Profiles_43.2 
#>   49     PAM_Profiles_44 
#>   50     PAM_Profiles_44.1 
#>   51     PAM_Profiles_44.2 
#>   52     PAM_Profiles_45 
#>   53     PAM_Profiles_45.1 
#>   54     PAM_Profiles_45.2 
#>   55     PAM_Profiles_46 
#>   56     PAM_Profiles_46.1 
#>   57     PAM_Profiles_46.2 
#>   58     PAM_WindDirection_28 
#>   59     PAM_WindDirection_28.1 
#>   60     PAM_WindDirection_28.2 
#>   61     PAM_WindDirection_29 
#>   62     PAM_WindDirection_29.1 
#>   63     PAM_WindDirection_29.2 
#>   64     PAM_WindDirection_30 
#>   65     PAM_WindDirection_30.1 
#>   66     PAM_WindDirection_30.2 
#>   67     PAM_WindDirection_31 
#>   68     PAM_WindDirection_31.1 
#>   69     PAM_WindDirection_31.2 
#>   70     PAM_WindDirection_32 
#>   71     PAM_WindDirection_32.1 
#>   72     PAM_WindDirection_32.2 
#>   73     PAM_WindDirection_33 
#>   74     PAM_WindDirection_33.1 
#>   75     PAM_WindDirection_33.2 
#>   76     PAM_WindDirection_34 
#>   77     PAM_WindDirection_34.1 
#>   78     PAM_WindDirection_34.2 
#>   79     PAM_WindDirection_35 
#>   80     PAM_WindDirection_35.1 
#>   81     PAM_WindDirection_35.2 
#>   82     PAM_WindDirection_36 
#>   83     PAM_WindDirection_36.1 
#>   84     PAM_WindDirection_36.2 
#>   85     PAM_WindDirection_37 
#>   86     PAM_WindDirection_37.1 
#>   87     PAM_WindDirection_37.2 
#>   88     PAM_WindDirection_38 
#>   89     PAM_WindDirection_38.1 
#>   90     PAM_WindDirection_38.2 
#>   91     PAM_WindDirection_39 
#>   92     PAM_WindDirection_39.1 
#>   93     PAM_WindDirection_39.2 
#>   94     PAM_WindDirection_40 
#>   95     PAM_WindDirection_40.1 
#>   96     PAM_WindDirection_40.2 
#>   97     PAM_WindDirection_41 
#>   98     PAM_WindDirection_41.1 
#>   99     PAM_WindDirection_41.2 
#>   100     PAM_WindDirection_42 
#>   101     PAM_WindDirection_42.1 
#>   102     PAM_WindDirection_42.2 
#>   103     PAM_WindDirection_43 
#>   104     PAM_WindDirection_43.1 
#>   105     PAM_WindDirection_43.2 
#>   106     PAM_WindDirection_44 
#>   107     PAM_WindDirection_44.1 
#>   108     PAM_WindDirection_44.2 
#>   109     PAM_WindDirection_45 
#>   110     PAM_WindDirection_45.1 
#>   111     PAM_WindDirection_45.2 
#>   112     PAM_WindDirection_46 
#>   113     PAM_WindDirection_46.1 
#>   114     PAM_WindDirection_46.2 
#> 
#> 
#> Additional notes:
#>   (*) You can access each layer using:
#>        1 - terra::subset function (See terra documentation)
#>        2 - the accessor syntax output[["nameOfLayer"]].
#>        3 - the accessor syntax output[[index]].
#>   (*) If any layers share the same name (e.g: same storm name), please use method 1 or 2.
plotBehaviour(pam, pam_prof[["PAM_Profiles_37"]], labels = TRUE, pos = 2)