Plot rasterized information storm behaviour

This function plots a rasterize product (maximum sustained wind, power dissipation index, category exposure, or 2D wind speed structure and Wind direction at a given observation) associated with a storm contained in a Storms object alongside with its track

plotBehaviour(
  sts,
  raster_product,
  color_palette = NULL,
  xlim = NULL,
  ylim = NULL,
  labels = FALSE,
  by = 8,
  pos = 3
)

Arguments

sts

Storms object

raster_product

Spatraster object. Name of the layer must be "stormName_product" where product is either "MSW", "PDI", "Exposure_min-max", "Profiles_index", "WindDirection_index" where "min" and "max" represent the threshold used to compute Exposure raster and index stands for the index of observation.

color_palette

character vector. Represents the color palette used for the plot. Default value is set to NULL, which will automatically choose a color palette provided by this package and depending on the product

xlim

numeric vector. A set of longitude coordinates that controls the longitude extent of the plot. Default value is set to NULL which will let the plot extends according to the x bounding box of the spatial LOI buffer provided in sts

ylim

numeric vector. A set of latitude coordinates that controls the latitude extent of the plot. Default value is set to NULL which will let the plot extends according to the y bounding box of the spatial LOI buffer provided in sts

labels

logical. Whether or not to plot ISO Times and name labels

by

numeric. Defines the frequency at which labels are plotted for the 3-hourly records. Default value is set to 8 which represents a 24h time interval between each labeled observations. Ignored if labels == FALSE

pos

numeric. Must be between 1 and 4. Correspond to the position of labels according to the observation: 1 (up), 2 (left), 3 (down), 4 (right). Default value is set to 3. Ignored if labels == FALSE

Examples

#Plot MSW analytic raster for Pam (2015) in Vanuatu
pam <- getStorms(loi = "Vanuatu", names = "PAM")
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching for PAM storm ...
#>    -> Identifying Storms: Done
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 0.6762528 sec ===
#> 
#> SUMMARY:
#> (*) LOI: Vanuatu 
#> (*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 1 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         PAM - 2015 - 5 - 20 
#> 
pam_msw <- stormBehaviour_sp(pam, verbose = 0)
plotBehaviour(pam, pam_msw)

#Plot a 2D windspeed structure  for Pam (2015) in Vanuatu
pam <- getStorms(loi = "Vanuatu", names = "PAM")
#> === getStorms processing ... ===
#> 
#> -> Making buffer: Done
#> -> Searching for PAM storm ...
#>    -> Identifying Storms: Done
#> -> Gathering storm(s) ... 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> === DONE with run time 0.675323 sec ===
#> 
#> SUMMARY:
#> (*) LOI: Vanuatu 
#> (*) Buffer size: 300 km
#> (*) Remove Tropical Depressions (< 18 m/s in sshs): yes
#> (*) Number of Storms: 1 
#>         Name - Tropical season - SSHS - Number of observation within buffer:
#>         PAM - 2015 - 5 - 20 
#> 
pam_prof <- stormBehaviour_sp(pam, product = "Profiles", verbose = 0)
plotBehaviour(pam, pam_prof[["PAM_Profiles_37"]], labels = TRUE, pos = 2)