Compute indicators of storm behaviour

The temporalBehaviour() function allows computing wind speed and direction for a given location or set of locations along the lifespan of a tropical cyclone. It also allows to compute to compute three associated summary statistics.

Usage

temporalBehaviour(
  sts,
  points,
  product = "TS",
  wind_threshold = c(18, 33, 42, 49, 58, 70),
  method = "Willoughby",
  asymmetry = "Chen",
  empirical_rmw = FALSE,
  temp_res = 1,
  verbose = 1
)

Arguments

sts

StormsList object.

points

data.frame. The lon (in a x col), lat (in a y col) coordinates in decimal degrees.

product

character. Desired output among:

"TS": time series of wind speed and direction (default setting).

"PDI": power dissipation index.

"Exposure": the duration of exposure.

wind_threshold

numeric vector. Minimal wind threshold(s) (in $m.s^-1$) used to compute the duration of exposure when product="Exposure". By default value the thresholds used in the Saffir-Simpson hurricane wind scale are used (i.e., 18, 33, 42, 49, 58, 70 $m.s^-1$).

method

character. Model used to compute wind speed and direction. Three different models are implemented:

"Willoughby": the symmetrical model developed by Willoughby et al. (2006) (default setting).

"Holland": the symmetrical model developed by Holland (1980).

"Boose": the asymmetrical model developed by Boose et al. (2004).

asymmetry

character. If method="Holland" or method="Willoughby", which method is used to add asymmetry. Can be:

"Chen": the model developed by Chen (1994) (default setting).

"Miyazaki": the model developed by Miyazaki et al. (1962).

"None": no asymmetry.

empirical_rmw

logical. Whether (TRUE) or not (FALSE) to compute the radius of maximum wind (rmw) empirically with the model developed by Willoughby et al. (2006). If empirical_rmw==FALSE (default setting) then the rmw provided in the StormsList is used.

temp_res

numeric. Temporal resolution. Can be 1 (for 60 min, default setting), 0.75 (for 45min), 0.5 (for 30 min), and 0.25 (15 for min).

verbose

numeric. Information displayed. Can be:

2: information about the processes and outputs are displayed (default setting).

1: information about the processes are displayed.

0: no information displayed.

Value

For each storm and each point location, the temporalBehaviour() function returns a data.frame. The data frames are organised into named lists. Depending on the product argument different data.frame are returned:

If product == "TS", the function returns a data.frame with one row for each observation (or interpolated observation) and four columns for wind speed (in $m.s^-1$), wind direction (in degree), the indices, and the ISO time of observations.

If product == "PDI", the function returns a data.frame with one row for each point location and one column for the PDI.

If product == "Exposure", the function returns a data.frame with one row for each wind speed threshold defined with the wind_threshold argument and one column for each point location.

Details

The temp_res input will perform a linear interpolation of observations to further compute each 2D wind speed structure at each interpolated observations. For example, if temp_res == 1, it will generate observations every 1 hour between the available observations.

References

Boose, E. R., Serrano, M. I., & Foster, D. R. (2004). Landscape and regional impacts of hurricanes in Puerto Rico. Ecological Monographs, 74(2), Article 2. https://doi.org/10.1890/02-4057

Chen, K.-M. (1994). A computation method for typhoon wind field. Tropic Oceanology, 13(2), 41–48.

Holland, G. J. (1980). An Analytic Model of the Wind and Pressure Profiles in Hurricanes. Monthly Weather Review, 108(8), 1212–1218. https://doi.org/10.1175/1520-0493(1980)108<1212:AAMOTW>2.0.CO;2

Miyazaki, M., Ueno, T., & Unoki, S. (1962). The theoretical investigations of typhoon surges along the Japanese coast (II). Oceanographical Magazine, 13(2), 103–117.

Willoughby, H. E., Darling, R. W. R., & Rahn, M. E. (2006). Parametric Representation of the Primary Hurricane Vortex. Part II: A New Family of Sectionally Continuous Profiles. Monthly Weather Review, 134(4), 1102–1120. https://doi.org/10.1175/MWR3106.1

Examples

if (FALSE) {
pts <- data.frame(x = c(166.5, 163), y = c(-22, -19))
row.names(pts)<-c("point_1","point_2")

#Computing time series of wind speed and direction for Erica and Niran over points 1 and 2 defined above
ts_nc <- temporalBehaviour(sts_nc, points = pts)

#Computing PDI for Erica and Niran over points 1 and 2 defined above
pdiPt_nc <- temporalBehaviour(sts_nc, points = pts, product = "PDI")

#Computing the duration of exposure to wind speeds above the thresholds used by the Saffir-Simpson hurricane wind scale for Erica and Niran over points 1 and 2 defined above
expPt_nc <- temporalBehaviour(sts_nc, points = pts, product = "Exposure", wind_threshold = c(20,30))
}